+++ ui/components/builder-header/toolbar/example.jsx
@@ -4,10 +4,6 @@
 import React from 'react';
 import BuilderHeader from '../';
 
-export const Context = props => (
-  <div style={{ position: 'relative', height: '100px' }}>{props.children}</div>
-);
-
 export default <BuilderHeader showToolbar />;
 
 export let states = [
@@ -16,5 +12,30 @@ export let states = [
     label: 'Menu open with toolbar',
     isOpen: true,
     element: <BuilderHeader openDropDown showToolbar />
+  },
+  {
+    id: 'has-status-text',
+    label: 'Has Status Text',
+    element: <BuilderHeader hasStatus showToolbar />
+  },
+  {
+    id: 'has-status-text-with-tooltip',
+    label: 'Has Status Text with Tooltip',
+    element: <BuilderHeader hasStatus hasStatusTooltip showToolbar />
+  },
+  {
+    id: 'has-status-text-with-warning',
+    label: 'Has Status Text with Warning',
+    element: <BuilderHeader hasStatus hasWarning showToolbar />
+  },
+  {
+    id: 'has-status-text-with-error',
+    label: 'Has Status Text with error',
+    element: <BuilderHeader hasStatus hasError showToolbar />
+  },
+  {
+    id: 'has-status-text-with-warning-and-error',
+    label: 'Has Status Text with warning and error',
+    element: <BuilderHeader hasStatus hasWarning hasError showToolbar />
   }
 ];