+++ ui/shared/empty-link/index.js
@@ -3,14 +3,15 @@ import PropTypes from 'prop-types';
 
 // Prevents React warning about javascript:void(0) usage
 // https://github.com/facebook/react/pull/15047
-const EmptyLink = props => (
-  <a href="#" onClick={e => e.preventDefault()}>
-    {props.children}
+const EmptyLink = ({ children, title }) => (
+  <a href="#" onClick={e => e.preventDefault()} title={title}>
+    {children}
   </a>
 );
 
 EmptyLink.propTypes = {
-  children: PropTypes.string
+  children: PropTypes.string,
+  title: PropTypes.string
 };
 
 export default EmptyLink;