+++ ui/components/builder-header/RELEASENOTES.md
@@ -4,6 +4,28 @@
 
 <!-- ## [Unreleased] -->
 
+## 2.15.0
+
+### Changed
+
+- Updates to top bar:
+  - Changed the color of the border separators to `$color-background-inverse`.
+  - On `:hover` and `:focus` for `.slds-builder-header__item-action`:
+    - Changed `background-color` to `$color-background-inverse`
+    - Changed `color` to `$color-gray-1`
+  - "Back" action moved to the left-hand side and visible text was removed
+  - "Link" renamed to "Settings" and moved to the right-hand side next to "Help"
+
+### Added
+
+- Additions to toolbar:
+  - Icons for canvas settings and toggling a panel
+  - Status text and alerts that convey status, warnings, and errors
+  - Popovers for alerts
+  - Tooltip for the "Back" action
+- New examples for "Back" action tooltip, status text, and alerts.
+- Added `aria-expanded` to dropdown item
+
 ## 2.14.0
 
 ### Fixed